+++ src/OpenTelemetry.Instrumentation.Wcf/Implementation/WcfInstrumentationActivitySource.cs
@@ -2,10 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Diagnostics;
-using System.Reflection;
 using System.ServiceModel.Channels;
 using OpenTelemetry.Instrumentation.Wcf.Implementation;
-using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Instrumentation.Wcf;
 
@@ -14,16 +12,17 @@ namespace OpenTelemetry.Instrumentation.Wcf;
 /// </summary>
 internal static class WcfInstrumentationActivitySource
 {
-    internal static readonly Assembly Assembly = typeof(WcfInstrumentationActivitySource).Assembly;
-    internal static readonly AssemblyName AssemblyName = Assembly.GetName();
-#pragma warning disable IDE0370 // Suppression is unnecessary
-    internal static readonly string ActivitySourceName = AssemblyName.Name!;
-#pragma warning restore IDE0370 // Suppression is unnecessary
-    internal static readonly string IncomingRequestActivityName = ActivitySourceName + ".IncomingRequest";
-    internal static readonly string OutgoingRequestActivityName = ActivitySourceName + ".OutgoingRequest";
-    internal static readonly string UnassociatedExceptionActivityName = ActivitySourceName + ".Exception";
+    internal static readonly Version SemanticConventionsVersion = new(1, 23, 0);
+    internal static readonly ActivitySource ActivitySource = Trace.ActivitySourceFactory.Create(typeof(WcfInstrumentationActivitySource), SemanticConventionsVersion);
 
-    public static ActivitySource ActivitySource { get; } = new(ActivitySourceName, Assembly.GetPackageVersion());
+    internal static readonly Version SemanticConventionsVersionNew = new(1, 42, 0);
+    internal static readonly ActivitySource ActivitySourceNew = Trace.ActivitySourceFactory.Create(typeof(WcfInstrumentationActivitySource), SemanticConventionsVersionNew);
+
+    internal static readonly ActivitySource ActivitySourceBoth = Trace.ActivitySourceFactory.Create(typeof(WcfInstrumentationActivitySource), null);
+
+    internal static readonly string IncomingRequestActivityName = ActivitySource.Name + ".IncomingRequest";
+    internal static readonly string OutgoingRequestActivityName = ActivitySource.Name + ".OutgoingRequest";
+    internal static readonly string UnassociatedExceptionActivityName = ActivitySource.Name + ".Exception";
 
     public static WcfInstrumentationOptions? Options { get; set; }
 
@@ -32,4 +31,9 @@ internal static class WcfInstrumentationActivitySource
 
     public static void MessageHeaderValueSetter(Message request, string name, string value)
         => TelemetryPropagationWriter.Default(request, name, value);
+
+    public static ActivitySource Get(WcfInstrumentationOptions? options) =>
+        options == null || !options.EmitNewRpcAttributes
+        ? ActivitySource
+        : options.EmitOldRpcAttributes ? ActivitySourceBoth : ActivitySourceNew;
 }