+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.16.0
+
+Released 2026-Jun-24
+
 * Fixed route template token matching so short route parameters no longer throw
   `IndexOutOfRangeException` during route extraction.
   ([#4340](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4340))