+++ src/OpenTelemetry.Instrumentation.GrpcCore/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Unreleased
 
-* **BREAKING**: Update to version 1.41.0 of the Semantic Conventions.
-  ([#4338](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4338))
+## 1.0.0-beta.12
+
+Released 2026-Jun-24
+
+* **BREAKING**: Update to version 1.42.0 of the Semantic Conventions.
+  ([#4338](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4338),
+  [#4508](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4508))
 
 * Add instrumentation scope version and schema URL to traces.
   ([#4338](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4338))