+++ src/OpenTelemetry.Instrumentation.GrpcNetClient/CHANGELOG.md
@@ -2,14 +2,23 @@
 
 ## Unreleased
 
-* **BREAKING**: Update to version 1.41.0 of the Semantic Conventions.
-  ([#4338](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4338))
+## 1.16.0-beta.1
+
+Released 2026-Jun-24
+
+* **BREAKING**: Update to version 1.42.0 of the Semantic Conventions.
+  ([#4338](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4338),
+  [#4508](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4508))
 
 * Add instrumentation scope version and schema URL to traces.
   ([#4338](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4338))
 
 * Updated OpenTelemetry core component version(s) to `1.16.0`.
   ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
+
+* Set the `error.type` attribute to the gRPC status code name on client spans when
+  the call fails.
+  ([#4508](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4508))
 
 ## 1.15.1-beta.1