+++ src/OpenTelemetry.Instrumentation.Owin/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.16.0-beta.1
+
+Released 2026-Jun-24
+
 * Update to version 1.41.0 of the HTTP Semantic Conventions.
   ([#4375](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4375))