+++ src/OpenTelemetry.Instrumentation.Wcf/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.16.0-beta.1
+
+Released 2026-Jun-24
+
 * Fixed an issue where server-side exception recording could bypass
   `IncomingRequestFilter` for filtered faulting requests.
   ([#4306](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4306))
@@ -20,6 +24,14 @@
 
 * Updated OpenTelemetry core component version(s) to `1.16.0`.
   ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
+
+* Add support for version [1.42.0](https://github.com/open-telemetry/semantic-conventions/blob/v1.42.0/docs/rpc/README.md)
+  of the Semantic Conventions for RPC when the `OTEL_SEMCONV_STABILITY_OPT_IN`
+  environment variable is set to `rpc` or `rpc/dup`.
+  ([#4377](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4377))
+
+* Add instrumentation scope version and schema URL to metrics and traces.
+  ([#4377](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4377))
 
 ## 1.15.1-beta.2